修改

+++ privilegeservice/src/main/java/cn/edu/xmu/privilegegateway/privilegeservice/model/bo/Privilege.java
@@ -89,7 +89,7 @@ public class Privilege implements VoObject{
      */
     private String cacuSignature;
 
-    private Byte sign;
+    private Integer sign;
 
     private Long creatorId;
